+++ include/geos/operation/buffer/OffsetCurveBuilder.h
@@ -40,6 +40,9 @@ class PrecisionModel;
 }
 }
 
+using geos::geom::CoordinateSequence;
+using geos::geom::PrecisionModel;
+
 namespace geos {
 namespace operation { // geos.operation
 namespace buffer { // geos.operation.buffer
@@ -55,8 +58,12 @@ namespace buffer { // geos.operation.buffer
  * it may contain self-intersections (and usually will).
  * The final buffer polygon is computed by forming a topological graph
  * of all the noded raw curves and tracing outside contours.
- * The points in the raw curve are rounded to a given geom::PrecisionModel.
+ * The points in the raw curve are rounded to a given PrecisionModel.
  *
+ * Note: this may not produce correct results if the input
+ * contains repeated or invalid points.
+ * Repeated points should be removed before calling.
+ * See removeRepeatedAndInvalidPoints.
  */
 class GEOS_DLL OffsetCurveBuilder {
 public:
@@ -68,13 +75,13 @@ public:
      *                   kept alive for the whole lifetime of
      *                   the buffer builder.
      */
-    OffsetCurveBuilder(const geom::PrecisionModel* newPrecisionModel,
-                       const BufferParameters& nBufParams)
-        :
-        distance(0.0),
-        precisionModel(newPrecisionModel),
-        bufParams(nBufParams)
-    {}
+    OffsetCurveBuilder(
+        const PrecisionModel* newPrecisionModel,
+        const BufferParameters& nBufParams)
+        : distance(0.0)
+        , precisionModel(newPrecisionModel)
+        , bufParams(nBufParams)
+        {}
 
     /** \brief
      * Gets the buffer parameters being used to generate the curve.
@@ -91,10 +98,9 @@ public:
      * Tests whether the offset curve for line or point geometries
      * at the given offset distance is empty (does not exist).
      * This is the case if:
-     * <ul>
-     * <li>the distance is zero,
-     * <li>the distance is negative, except for the case of singled-sided buffers
-     * </ul>
+     *
+     *  * the distance is zero,
+     *  * the distance is negative, except for the case of singled-sided buffers
      *
      * @param distance the offset curve distance
      * @return true if the offset curve is empty
@@ -113,9 +119,23 @@ public:
      *                 CoordinateSequences will be pushed_back.
      *                 Caller is responsible to delete these new elements.
      */
-    void getLineCurve(const geom::CoordinateSequence* inputPts,
-                      double distance,
-                      std::vector<geom::CoordinateSequence*>& lineList);
+    void getLineCurve(const CoordinateSequence* inputPts,
+        double distance,
+        std::vector<CoordinateSequence*>& lineList);
+
+    /**
+    * This method handles single points as well as LineStrings.
+    * LineStrings are assumed <b>not</b> to be closed (the function will not
+    * fail for closed lines, but will generate superfluous line caps).
+    *
+    * @param inputPts the vertices of the line to offset
+    * @param pDistance the offset distance
+    *
+    * @return a Coordinate array representing the curve
+    * or null if the curve is empty
+    */
+    std::unique_ptr<CoordinateSequence> getLineCurve(
+        const CoordinateSequence* inputPts, double pDistance);
 
     /** \brief
      * This method handles single points as well as lines.
@@ -135,8 +155,8 @@ public:
      *
      * @note This is a GEOS extension.
      */
-    void getSingleSidedLineCurve(const geom::CoordinateSequence* inputPts,
-                                 double distance, std::vector<geom::CoordinateSequence*>& lineList,
+    void getSingleSidedLineCurve(const CoordinateSequence* inputPts,
+                                 double distance, std::vector<CoordinateSequence*>& lineList,
                                  bool leftSide, bool rightSide) ;
 
     /** \brief
@@ -149,19 +169,38 @@ public:
      * @param lineList the std::vector to which CoordinateSequences will
      *                 be pushed_back
      */
-    void getRingCurve(const geom::CoordinateSequence* inputPts, int side,
+    void getRingCurve(const CoordinateSequence* inputPts, int side,
                       double distance,
-                      std::vector<geom::CoordinateSequence*>& lineList);
+                      std::vector<CoordinateSequence*>& lineList);
 
-    void getOffsetCurve(const geom::CoordinateSequence* inputPts,
+    /**
+    * This method handles the degenerate cases of single points and lines,
+    * as well as valid rings.
+    *
+    * @param inputPts the coordinates of the ring (must not contain repeated points)
+    * @param side side the side Position of the ring on which to construct the buffer line
+    * @param pDistance the positive distance at which to create the offset
+    * @return a Coordinate array representing the curve,
+    * or null if the curve is empty
+    */
+    std::unique_ptr<CoordinateSequence> getRingCurve(
+        const CoordinateSequence* inputPts,
+        int side, double pDistance);
+
+    void getOffsetCurve(const CoordinateSequence* inputPts,
                         double p_distance,
-                        std::vector<geom::CoordinateSequence*>& lineList);
+                        std::vector<CoordinateSequence*>& lineList);
+
+    std::unique_ptr<CoordinateSequence> getOffsetCurve(
+        const CoordinateSequence* inputPts,
+        double pDistance);
+
 
 private:
 
     double distance;
 
-    const geom::PrecisionModel* precisionModel;
+    const PrecisionModel* precisionModel;
 
     const BufferParameters& bufParams;
 
@@ -183,20 +222,23 @@ private:
      */
     double simplifyTolerance(double bufDistance);
 
-    void computeLineBufferCurve(const geom::CoordinateSequence& inputPts,
+    void computeLineBufferCurve(const CoordinateSequence& inputPts,
                                 OffsetSegmentGenerator& segGen);
 
-    void computeSingleSidedBufferCurve(const geom::CoordinateSequence& inputPts,
+    void computeSingleSidedBufferCurve(const CoordinateSequence& inputPts,
                                        bool isRightSide,
                                        OffsetSegmentGenerator& segGen);
 
-    void computeRingBufferCurve(const geom::CoordinateSequence& inputPts,
+    void computeRingBufferCurve(const CoordinateSequence& inputPts,
                                 int side, OffsetSegmentGenerator& segGen);
-
-    std::unique_ptr<OffsetSegmentGenerator> getSegGen(double dist);
 
     void computePointCurve(const geom::Coordinate& pt,
                            OffsetSegmentGenerator& segGen);
+
+    void computeOffsetCurve(
+        const CoordinateSequence* inputPts,
+        bool isRightSide,
+        OffsetSegmentGenerator& segGen);