+++ NEWS.md
@@ -51,7 +51,7 @@ xxxx-xx-xx
   - HausdorffDistance: Fix crash on collection containing empty point (GH-840, Dan Baston)
   - MaximumInscribedCircle: Fix infinite loop with non-finite coordinates (GH-843, Dan Baston)
   - DistanceOp: Fix crash on collection containing empty point (GH-842, Dan Baston)
-
+  - Improve OffsetCurve behaviour and add Joined mode (JTS-956, Martin Davis)
 
 ## Changes in 3.11.0
 2022-07-01